requirement.txt
streamlit==1.29.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.0
uuid